Manage User Roles And Permissions

+++ src/io/reflection/app/client/controller/SessionController.java
@@ -331,12 +331,12 @@ public class SessionController implements ServiceConstants, JsonServiceCallEvent
 	 * @param id
 	 * @return Boolean hasRole
 	 */
-	public boolean hasRole(User user, long id) {
+	public boolean hasRole(User user, Long id) {
 		boolean hasRole = false;
 
 		if (user != null && user.roles != null) {
 			for (Role role : user.roles) {
-				if (role.id != null && role.id.longValue() == id) {
+				if (role.id != null && role.id.longValue() == id.longValue()) {
 					hasRole = true;
 					break;
 				}
@@ -346,14 +346,14 @@ public class SessionController implements ServiceConstants, JsonServiceCallEvent
 		return hasRole;
 	}
 
-	public boolean hasPermission(User user, long id) {
+	public boolean hasPermission(User user, Long id) {
 		boolean hasPermission = hasRole(user, DataTypeHelper.ROLE_ADMIN_ID);
 
 		if (!hasPermission && user != null) {
 			if (user.roles != null) {
 				for (Role role : user.roles) {
 					if (!hasPermission && role.permissions != null) {
-						hasPermission = hasIdPermission(role.permissions, id);
+						hasPermission = hasIdPermission(role.permissions, id.longValue());
 						if (hasPermission) {
 							break;
 						}
@@ -361,7 +361,7 @@ public class SessionController implements ServiceConstants, JsonServiceCallEvent
 				}
 			}
 			if (!hasPermission && user.permissions != null) {
-				hasPermission = hasIdPermission(user.permissions, id);
+				hasPermission = hasIdPermission(user.permissions, id.longValue());
 			}
 
 		}
